Predicates added inside core game loop.
Changes:
Added a new struct data type "userOutcomedata" which keeps track of the users score

added a copy_if iterator which returns any characters in the user's input which share the same character and index as the answer

added a find_if to return any characters that are the same but dont share an index.

renamed WordCharacterData to AnswerCharacterData

Removed some data from AnswerCharacterData to userOutcomeData

Edited:

Added a colour scheme based on the user's score.